fix: Move `AccessScope` onto it's own path
Move from:

```
import { AccessScope } from "@guardian/cdk/lib/patterns/ec2-app";
```

To:

```
import { AccessScope } from "@guardian/cdk/lib/constants/access";
```
